Safari-611.1.10/LayoutTests/http/tests/cache/disk-cache/disk-cache-cancel-expected.txt

Test that canceled network loads don't leave behind stale cache entries

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 3 tests

Warming cache
Starting loads, then canceling
Loading again with back navigation cache policy
response headers: {"Cache-control":"max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
response source: Network

response headers: {"Cache-control":"max-age=0","ETag":"match"}
response source: Network

PASS successfullyParsed is true

TEST COMPLETE
